smarty and module fx changes
