+++ config.json
@@ -14,7 +14,7 @@
       "institution": "Marshfield Clinic Research Institute"
     }
   ],
-  "framework-version": 6,
+  "framework-version": 12,
   "links": {
     "project": [
       {
@@ -25,9 +25,6 @@
       }
     ]
   },
-  "permissions": [
-    "redcap_every_page_top"
-  ],
   "project-settings": [
     {
       "key": "replace_add_edit",
